nodeElementsをコンポーネントに分離。
